pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "rdjat-embed"
version = "1.0.0"
description = "A GUI application for image steganography using the RDJAT average-bin method for embedding and extraction"
authors = [
    {name = "RDJAT-Embed Project"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Security :: Cryptography",
]
keywords = ["steganography", "image-processing", "gui", "embedding", "extraction", "RDJAT", "RDJAT-Embed"]
dependencies = [
    "numpy>=1.24,<3.0",
    "Pillow>=10,<12",
    "scikit-image>=0.22,<0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[project.scripts]
rdjat-embed = "rdjat:main"

[project.urls]
Homepage = ""
Repository = ""
Documentation = ""
"Bug Reports" = ""

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=rdjat --cov-report=html --cov-report=term-missing"